+++ src/app/officers/officers-dashboard/officers-dashboard.component.html
@@ -17,12 +17,7 @@
   <p-dataview #dv [value]="officers()" *ngIf="!isLoading">
     <ng-template #list let-items>
       <div class="grid grid-cols-12 gap-0 p-4">
-        <div
-          class="col-span-3 listItem"
-          *ngFor="let item of items; let first = first"
-          routerLink="/flight"
-          [queryParams]="{ id: item.id }"
-        >
+        <div class="col-span-3 listItem" *ngFor="let item of items">
           <div class="flex flex-col sm:flex-row sm:items-center p-2 gap-4">
             <div
               class="flex flex-col md:flex-row justify-between md:items-center flex-1 gap-6 ml-4"
