+++ src/app/tournaments/tournaments-dashboard/tournaments-dashboard.component.css
